Added retchmail, which is probably the world's fastest POP client what with the way it doesn't follow any standards.  But it sure is fast.
WvStream::getline() now uses continue_select() if uses_continue_select is
true for that stream.  We'll see how much stuff that breaks... also, added a
terminate_continue_select() that you're supposed to call in your destructor
to clean things up if you use continue_select.  Because destructors are
almost as stupid as constructors, I can't do this for you in
WvStream::~WvStream().


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@343 6619c942-59fa-0310-bf43-a07aaa81d9ac